chore: move Garden Cloud utils to own module
