+++ core/src/cloud/api.ts
@@ -26,7 +26,8 @@ import type {
   GetProjectResponse,
   ListProjectsResponse,
 } from "@garden-io/platform-api-types"
-import { getCloudDistributionName, getCloudLogSectionName, getPackageVersion } from "../util/util.js"
+import { getCloudDistributionName, getCloudLogSectionName } from "../util/garden-cloud.js"
+import { getPackageVersion } from "../util/util.js"
 import type { CommandInfo } from "../plugin-context.js"
 import type { ClientAuthToken, GlobalConfigStore } from "../config-store/global.js"
 import { add } from "date-fns"
