+++ core/src/commands/cloud/secrets/secrets-update.ts
@@ -13,8 +13,8 @@ import type { CloudProject } from "../../../cloud/api.js"
 import type { StringMap } from "../../../config/common.js"
 import { CloudApiError, CommandError, ConfigurationError, GardenError } from "../../../exceptions.js"
 import { printHeader } from "../../../logger/util.js"
+import { getCloudDistributionName } from "../../../util/garden-cloud.js"
 import { dedent, deline } from "../../../util/string.js"
-import { getCloudDistributionName } from "../../../util/util.js"
 import type { CommandParams, CommandResult } from "../../base.js"
 import { Command } from "../../base.js"
 import type { ApiCommandError, SecretResult } from "../helpers.js"
