+++ core/src/commands/logout.ts
@@ -10,8 +10,8 @@ import type { CommandParams, CommandResult } from "./base.js"
 import { Command } from "./base.js"
 import { printHeader } from "../logger/util.js"
 import { CloudApi, getGardenCloudDomain } from "../cloud/api.js"
+import { getCloudDistributionName } from "../util/garden-cloud.js"
 import { dedent, deline } from "../util/string.js"
-import { getCloudDistributionName } from "../util/util.js"
 import { ConfigurationError } from "../exceptions.js"
 import type { ProjectConfig } from "../config/project.js"
 import { findProjectConfig } from "../config/base.js"
