+++ core/src/commands/serve.ts
@@ -14,7 +14,8 @@ import { printEmoji, printHeader } from "../logger/util.js"
 import { dedent } from "../util/string.js"
 import type { CommandLine } from "../cli/command-line.js"
 import { GardenInstanceManager } from "../server/instance-manager.js"
-import { getCloudDistributionName, sleep } from "../util/util.js"
+import { getCloudDistributionName } from "../util/garden-cloud.js"
+import { sleep } from "../util/util.js"
 import type { Log } from "../logger/log-entry.js"
 import { findProjectConfig } from "../config/base.js"
 import { CloudApiTokenRefreshError, getGardenCloudDomain } from "../cloud/api.js"
